backend/requirements.txt
# Django
Django>=6.0,<7.0  # :contentReference[oaicite:3]{index=3}
djangorestframework>=3.16.1  # latest stable :contentReference[oaicite:4]{index=4}
django-cors-headers>=4.9.0  # latest stable :contentReference[oaicite:5]{index=5}
djangorestframework-simplejwt>=5.3.1

# Agents & LLMs
pydantic>=2.5.2
openai>=2.11.0  # latest stable :contentReference[oaicite:6]{index=6}

langchain==1.1.3  # :contentReference[oaicite:7]{index=7}
langchain-core==1.1.3  # :contentReference[oaicite:8]{index=8}
langchain-community>=0.4.1,<1.0.0  # latest line :contentReference[oaicite:9]{index=9}
langchain-text-splitters>=1.0.0,<2.0.0  # :contentReference[oaicite:10]{index=10}

langchain-openai>=1.1.2,<2.0.0  # :contentReference[oaicite:11]{index=11}
langchain-google-genai>=4.0.0,<5.0.0  # uses google-genai SDK :contentReference[oaicite:12]{index=12}
langchain-anthropic>=1.2.0,<2.0.0  # :contentReference[oaicite:13]{index=13}
langchain-groq>=1.1.0,<2.0.0  # :contentReference[oaicite:14]{index=14}
langchain-huggingface>=1.1.0,<2.0.0  # :contentReference[oaicite:15]{index=15}
langchain-ollama>=1.0.0,<2.0.0  # :contentReference[oaicite:16]{index=16}

# Vectorstores & Embeddings
faiss-cpu>=1.7.4
sentence-transformers>=2.6.0

# Document loaders
python-docx>=1.1.0
python-multipart>=0.0.9
pypdf>=4.2.0
docx2txt>=0.8

# Deployment
gunicorn>=21.2.0
requests>=2.31.0
